2010-08-09	Ralf Corsépius <ralf.corsepius@rtems.org>
	* posix/include/aio.h, posix/src/aio_return.c:
	Let aio_return return ssize_t (Mandated by POSIX).

+++ cpukit/posix/src/aio_return.c
@@ -22,7 +22,7 @@
 #include <rtems/system.h>
 #include <rtems/seterr.h>
 
-int aio_return(
+ssize_t aio_return(
   const struct aiocb  *aiocbp __attribute__((unused))
 )
 {
